requirements.txt
pynput==1.7.7
mss==9.0.1
Pillow==10.4.0
google-generativeai==0.8.3
aiogram==3.15.0
python-dotenv==1.0.1